+++ paw-webapp/webapp/pom.xml
@@ -157,11 +157,6 @@
             <groupId>org.flywaydb</groupId>
             <artifactId>flyway-core</artifactId>
         </dependency>
-        <!-- EMails -->
-        <dependency>
-            <groupId>javax.mail</groupId>
-            <artifactId>mail</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
